Refactor: Prize enum을 Constant 패키지로 이동
